Refactor blog
Extract out BlogPost and BlogSidebar as own files. This is going to be useful in order to build the blog listing page
